src/main/python/requirements.txt
beautifulsoup4==4.10.0
pymupdf
requests==2.21.0
summa==1.2.0
python-dotenv==1.0.1
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0